Reduce CUDA hash calculation overhead by a factor of 4

+++ libethash-cuda/CUDAMiner.h
@@ -76,7 +76,7 @@ private:
     cudaStream_t* m_streams = nullptr;
     uint64_t m_current_target = 0;
 
-    uint8_t m_searchPasses = 0;
+    uint16_t m_searchPasses = 0;
 
     /// The local work size for the search
     static unsigned s_blockSize;
